feat: relocated styled folder and imports
